Moves player and tests files to Gameplay namespace
