+++ Assets/Prefabs/Player.prefab
@@ -11,11 +11,11 @@ GameObject:
   - component: {fileID: 817833974993414891}
   - component: {fileID: 8887210594514678503}
   - component: {fileID: -8230587324157615285}
-  - component: {fileID: 1258119580588674689}
   - component: {fileID: 2597257185044323647}
+  - component: {fileID: 1481507867024084348}
   m_Layer: 0
   m_Name: Player
-  m_TagString: Untagged
+  m_TagString: Player
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
@@ -108,18 +108,6 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 0
   m_Constraints: 0
---- !u!114 &1258119580588674689
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3976692843403402619}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d795416e8a221cd4b86df8603a610d8c, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!61 &2597257185044323647
 BoxCollider2D:
   m_ObjectHideFlags: 0
@@ -146,3 +134,16 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 0.48, y: 0.48}
   m_EdgeRadius: 0
+--- !u!114 &1481507867024084348
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3976692843403402619}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 59b13c9bc395f3947a46e19c8fccfe8c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  speed: 0
