+++ Assets/Scripts/Gameplay/IPlayerInput.cs.meta
@@ -1,3 +1,3 @@
 ﻿fileFormatVersion: 2
-guid: ec676a095a2b454a86650e00e231d08b
+guid: 6de2bb2ecbef80e4eb7015bc23a93d70
 timeCreated: 1650113044
