+++ Assets/Scripts/Gameplay/Player.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2641428d807b19b47bc5c1db022be0b3
+guid: 59b13c9bc395f3947a46e19c8fccfe8c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
