--- Assets/Scripts/Gameplay/Player.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 83d33b0729e295444833dd555b346c59
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
