--- Assets/Scripts/Gameplay/Player/Player.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d795416e8a221cd4b86df8603a610d8c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
