--- Assets/Tests/Gameplay/Player.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 937e45d8fb401bc4c90d5ca70b982660
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
